backend/pyproject.toml
[tool.poetry]
name = "xano-logs"
version = "0.1.0"
description = ""
authors = ["Parth Pisat <parth.pisat@unicoconnect.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.2.2"
pandas = "^2.2.3"
seaborn = "^0.13.2"
ollama = "^0.4.7"
requests = "^2.32.3"
jira = "^3.8.0"
slack-sdk = "^3.34.0"
aiohttp = "^3.11.11"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
pydantic = "^2.10.6"
watchdog = "^6.0.0"
python-dotenv = "^1.0.1"
streamlit = "^1.42.0"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
